+++ xfinity-usage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.0.12.7
+
+- Minor code cleanup
+- When configured for Debug logging, a log file is now created in the addons_config folder
+- Two Step verification and Akamai Access Denied detection
+- Browser persistent storage is now used
+- The script removes persistent storage and xfinity.log from the addon_config folder during startup
+- The Firefox user agent and version is not randomized
+
 ## 0.0.12.6
 
 - Switched to Debian 12 docker
@@ -22,11 +31,11 @@
 
 ## 0.0.11.1
 
-- Resolved issue with xFinity customers who have the unlimited plan and/or are using the xFi Gateway
+- Resolved issue with Xfinity customers who have the unlimited plan and/or are using the xFi Gateway
 
 ## 0.0.11.0.4
 
-- Working on fixing issue with Unlimited Data Plan and xFinity Wifi modem
+- Working on fixing issue with Unlimited Data Plan and Xfinity Wifi modem
 
 ## 0.0.11
 
@@ -45,7 +54,7 @@
 - Added a lot more page checks to make sure page is properly loaded
 - Session data is now based on the OAuth JWT token
 - Added PyJWT library to Docker file to allow for JWT token processing
-- Encypt the Xfinity password in the log file when in debug mode
+- Encrypt the Xfinity password in the log file when in debug mode
 - Added debug_support logging option to gather page source and screenshots when extra troubleshooting is needed.
 - Version bump pip requirements:
     - playwright to 1.44.0